--- src/Concerns/CanTransition.php
@@ -1,67 +1,0 @@
-<?php
-
-namespace OwenIt\Auditing\Concerns;
-
-use OwenIt\Auditing\Contracts\AttributeRedactor;
-use OwenIt\Auditing\Contracts\Audit as AuditContract;
-use OwenIt\Auditing\Contracts\Auditable as AuditableContract;
-use OwenIt\Auditing\Exceptions\AuditableTransitionException;
-
-trait CanTransition
-{
-    /**
-     * {@inheritdoc}
-     */
-    public function transitionTo(AuditContract $audit, bool $old = false): AuditableContract
-    {
-        // The Audit must be for an Auditable model of this type
-        if ($this->getMorphClass() !== $audit->auditable_type) {
-            throw new AuditableTransitionException(sprintf(
-                'Expected Auditable type %s, got %s instead',
-                $this->getMorphClass(),
-                $audit->auditable_type
-            ));
-        }
-
-        // The Audit must be for this specific Auditable model
-        if ($this->getKey() !== $audit->auditable_id) {
-            throw new AuditableTransitionException(sprintf(
-                'Expected Auditable id (%s)%s, got (%s)%s instead',
-                gettype($this->getKey()),
-                $this->getKey(),
-                gettype($audit->auditable_id),
-                $audit->auditable_id
-            ));
-        }
-
-        // Redacted data should not be used when transitioning states
-        foreach ($this->getAttributeModifiers() as $attribute => $modifier) {
-            if (is_subclass_of($modifier, AttributeRedactor::class)) {
-                throw new AuditableTransitionException('Cannot transition states when an AttributeRedactor is set');
-            }
-        }
-
-        // The attribute compatibility between the Audit and the Auditable model must be met
-        $modified = $audit->getModified();
-
-        if ($incompatibilities = array_diff_key($modified, $this->getAttributes())) {
-            throw new AuditableTransitionException(sprintf(
-                'Incompatibility between [%s:%s] and [%s:%s]',
-                $this->getMorphClass(),
-                $this->getKey(),
-                get_class($audit),
-                $audit->getKey()
-            ), array_keys($incompatibilities));
-        }
-
-        $key = $old ? 'old' : 'new';
-
-        foreach ($modified as $attribute => $value) {
-            if (array_key_exists($key, $value)) {
-                $this->setAttribute($attribute, $value[$key]);
-            }
-        }
-
-        return $this;
-    }
-}
